Take selected tags out of the search bar
