+++ client-v2/src/components/TextInput.js
@@ -2,9 +2,9 @@
 
 import * as React from 'react';
 import styled from 'styled-components';
-
 import moreImage from 'shared/images/icons/more.svg';
 import searchImage from 'shared/images/icons/search.svg';
+import { SmallRoundButton, ClearButtonIcon } from 'util/sharedStyles/buttons';
 
 const InputWrapper = styled('div')`
   position: relative;
@@ -12,20 +12,6 @@ const InputWrapper = styled('div')`
   display: flex;
   border: solid 1px #c9c9c9;
   backgroud: #fffff;
-`;
-
-const TagItem = styled('div')`
-  color: #121212;
-  font-weight: bold;
-  border: solid 1px #c4c4c4;
-  font-size: 12px;
-  background-color: #ffffff;
-  padding: 7px 15px 7px 15px;
-  border-top: solid 1px #121212;
-  margin: 5px;
-  :hover {
-    color: #c4c4c4;
-  }
 `;
 
 const Input = styled(`input`)`
@@ -43,31 +29,6 @@ const Input = styled(`input`)`
 
   &::placeholder
     color: rgba(255, 255, 255, 0.75);
-  }
-`;
-
-const SmallRoundButton = styled('button')`
-  appearance: none;
-  display: inline-block;
-  vertical-align: middle;
-  text-align: center;
-  width: 32px;
-  height: 32px;
-  border: none;
-  color: inherit;
-  cursor: pointer;
-  padding: 0;
-  border-radius: 100%;
-  transition: background-color 0.15s;
-
-  ::before {
-    font-size: 1em;
-    line-height: 1;
-  }
-
-  :focus {
-    outline: none;
-    font-weight: bold;
   }
 `;
 
@@ -92,12 +53,6 @@ const ButtonsContainer = styled('div')`
   right: 8px;
 `;
 
-const ClearButtonIcon = styled('img')`
-  transform: rotate(45deg);
-  vertical-align: middle;
-  margin-right: 4px;
-`;
-
 const SearchButtonIcon = styled('img')`
   vertical-align: middle;
 `;
@@ -107,9 +62,7 @@ type TextInputProps = {
   onSearch?: () => void,
   onDisplaySearchFilters?: () => void,
   width?: string,
-  searchTermsExist: boolean,
-  searchTerms: Array<string>,
-  onClearTag: string => void
+  searchTermsExist: boolean
 };
 
 const TextInput = ({
@@ -117,28 +70,9 @@ const TextInput = ({
   onSearch,
   searchTermsExist,
   onDisplaySearchFilters,
-  searchTerms,
-  onClearTag,
   ...props
 }: TextInputProps) => (
   <InputWrapper>
-    {searchTerms.map(searchTerm => (
-      <TagItem key={searchTerm}>
-        <span>{searchTerm}</span>
-        <SmallRoundButton
-          onClick={() => onClearTag(searchTerm)}
-          title="Clear search"
-        >
-          <ClearButtonIcon
-            src={moreImage}
-            onClick={() => onClearTag(searchTerm)}
-            alt=""
-            height="22px"
-            width="22px"
-          />
-        </SmallRoundButton>
-      </TagItem>
-    ))}
     <Input {...props} />
     <ButtonsContainer>
       {onClear &&
